Fix types for macos

+++ application/src/server/filesystem/cap/mod.rs
@@ -764,7 +764,7 @@ impl CapFilesystem {
                 rustix::fs::chmodat(
                     inner.as_fd(),
                     path,
-                    rustix::fs::Mode::from_raw_mode(permissions.mode),
+                    rustix::fs::Mode::from_raw_mode(permissions.mode as u16),
                     rustix::fs::AtFlags::SYMLINK_NOFOLLOW,
                 )
             })
@@ -801,7 +801,7 @@ impl CapFilesystem {
                 rustix::fs::chmodat(
                     inner.as_fd(),
                     path,
-                    rustix::fs::Mode::from_raw_mode(permissions.mode),
+                    rustix::fs::Mode::from_raw_mode(permissions.mode as u16),
                     rustix::fs::AtFlags::SYMLINK_NOFOLLOW,
                 )?;
             }
